requirements.txt
Django>=4.2.29
djangorestframework>=3.16,<3.17
psycopg[binary]>=3.1,<4
